backend/requirements.txt
flask_sqlalchemy==3.1.1
SQLAlchemy==2.0.41
